--- app/controllers/api/JacksonRepresentation.java
@@ -1,55 +1,0 @@
-package controllers.api;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import play.mvc.Result;
-import uk.gov.openregister.config.Register;
-import uk.gov.openregister.domain.Record;
-import uk.gov.openregister.domain.RecordVersionInfo;
-
-import java.util.List;
-import java.util.Map;
-
-import static play.mvc.Results.ok;
-
-public class JacksonRepresentation implements Representation {
-    private final ObjectMapper objectMapper;
-    private final String contentType;
-
-    public JacksonRepresentation(ObjectMapper objectMapper, String contentType) {
-        this.objectMapper = objectMapper;
-        this.contentType = contentType;
-    }
-
-    @Override
-    public Result toListOfRecords(Register register,
-                                  List<Record> records,
-                                  Map<String, String[]> requestParams,
-                                  Map<String, String> representationsMap,
-                                  String previousPageLink,
-                                  String nextPageLink) {
-        return ok(asString(requestParams, records)).as(contentType);
-    }
-
-    @Override
-    public Result toRecord(Register register,
-                           Record record,
-                           Map<String, String[]> requestParams,
-                           Map<String, String> representationsMap,
-                           List<RecordVersionInfo> history) {
-        return ok(asString(requestParams, record)).as(contentType);
-    }
-
-    @Override
-    public boolean isPaginated() {
-        return false;
-    }
-
-    protected String asString(Map<String, String[]> requestParams, Object record) {
-        try {
-            return objectMapper.writeValueAsString(record);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
