+++ app/controllers/api/JsonRepresentation.java
@@ -1,29 +1,25 @@
 package controllers.api;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
 import play.mvc.Result;
+import play.mvc.Results;
+import uk.gov.openregister.JsonObjectMapper;
+import uk.gov.openregister.config.Register;
+import uk.gov.openregister.domain.Record;
+import uk.gov.openregister.domain.RecordVersionInfo;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
+import static play.mvc.Results.ok;
 import static play.mvc.Results.status;
 
-public class JsonRepresentation extends JacksonRepresentation {
+public class JsonRepresentation implements Representation {
 
-    public static final String CONTENT_TYPE = "application/json; charset=utf-8";
-
-    private JsonRepresentation() {
-        super(makeObjectMapper(), CONTENT_TYPE);
-    }
-
-    private static ObjectMapper makeObjectMapper() {
-        ObjectMapper objectMapper = new ObjectMapper();
-        objectMapper.configure(SerializationFeature.ORDER_MAP_ENTRIES_BY_KEYS, true);
-        return objectMapper;
-    }
+    public static final String JSON_CONTENT_TYPE = "application/json; charset=utf-8";
+    public static final String JSONP_CONTENT_TYPE = "application/javascript; charset=utf-8";
 
     public Result createdResponse() {
         Map<String, Object> result = new HashMap<>();
@@ -31,17 +27,42 @@ public class JsonRepresentation extends JacksonRepresentation {
         result.put("message", "Record saved successfully");
         result.put("errors", emptyList());
 
-        return status(202, super.asString(emptyMap(), result)).as(CONTENT_TYPE);
+        return status(202, asString(emptyMap(), result)).as(JSON_CONTENT_TYPE);
+    }
+
+    public static JsonRepresentation instance = new JsonRepresentation();
+
+    @Override
+    public Result toListOfRecords(Register register,
+                                  List<Record> records,
+                                  Map<String, String[]> requestParams,
+                                  Map<String, String> representationsMap,
+                                  String previousPageLink,
+                                  String nextPageLink) {
+        Results.Status ok = ok(asString(requestParams, records));
+        return requestParams.get("_callback") != null ? ok.as(JSONP_CONTENT_TYPE) : ok.as(JSON_CONTENT_TYPE);
     }
 
     @Override
+    public Result toRecord(Register register,
+                           Record record,
+                           Map<String, String[]> requestParams,
+                           Map<String, String> representationsMap,
+                           List<RecordVersionInfo> history) {
+        Results.Status ok = ok(asString(requestParams, record));
+        return requestParams.get("_callback") != null ? ok.as(JSONP_CONTENT_TYPE) : ok.as(JSON_CONTENT_TYPE);
+    }
+
+    @Override
+    public boolean isPaginated() {
+        return false;
+    }
+
     protected String asString(Map<String, String[]> requestParams, Object record) {
-        String json = super.asString(requestParams, record);
+        String json = JsonObjectMapper.convertToString(record);
 
         String[] callbacks = requestParams.get("_callback");
 
         return callbacks != null ? callbacks[0] + "(" + json + ");" : json;
     }
-
-    public static JsonRepresentation instance = new JsonRepresentation();
 }
